% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balloon_h.R
\name{bw_balloon_LQ}
\alias{bw_balloon_LQ}
\title{Loftsgaarden-Quesenberry kth nearest neighbor bandwidths}
\usage{
bw_balloon_LQ(tdat, edat, k)
}
\arguments{
\item{tdat}{vector or array containing the training dataset.}

\item{edat}{vector or array containing the evaluation data points.}

\item{k}{integer specifying the order of the neighbor.}
}
\value{
a vector of bandwidths  (one bandwidth per evaluation data point)
}
\description{
\code{bw_balloon_LQ} looks for the bandwidth following the
Loftsgaarden-Quesenberry algorithm. The bandwidth depends on the point of
estimation.
}
\details{
These are the details of the function.
}
\examples{
# univariate
bw_balloon_LQ(tdat = 1:10,edat = 5.1,k = 3)
bw_balloon_LQ(tdat = 1:10,edat = c(2.8,5.1),k = 3)
# multivariate
bw_balloon_LQ(tdat = array(1:20,dim=c(4,5)),edat = 1:4,k = 1)
}

