% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSmoothing_h.R
\name{bw_sampleSmoothing}
\alias{bw_sampleSmoothing}
\title{Loftsgaarden-Quesenberry kth nearest neighbor bandwidths}
\usage{
bw_sampleSmoothing(tdat, k)
}
\arguments{
\item{tdat}{a vector containing the training dataset.}

\item{k}{specifying the order of the neighbor.}
}
\value{
a vector of bandwidths  (one bandwidth per evaluation data point)
}
\description{
\code{bw_balloon_LQ} looks for the bandwidth following the
Loftsgaarden-Quesenberry algorithm. The bandwidth depends on the point of
estimation.
}
\details{
These are the details of the function.
}
\examples{
bw_sampleSmoothing(tdat = 1:10,k = 3)
bw_sampleSmoothing(tdat = (1:10)^2,k = 3)
}

