% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vkde.R
\name{kde_h}
\alias{kde_h}
\title{title.}
\usage{
kde_h(tdat, edat, bwtype, h)
}
\arguments{
\item{tdat}{a vector containing the training dataset.}

\item{edat}{a vector containing the evaluation data points.}

\item{bwtype}{a string corresponding to the method for calculating
bandwidths. Either \code{'fixed'}, \code{'balloon'} or
\code{'sampleSmoothing'}.}

\item{h}{the bandwdth(s) of the kernels. It should be a scalar when
\code{'bwtype=fixed'}, a vector of length \code{length(edat)} for
\code{'bwtype='balloon'}, a vector of length \code{length(tdat)} for
\code{'bwtype='sampleSmoothing'}. It can be calculated from other functions
in this package.}
}
\value{
The density at locations \code{edat} based on the training dataset
  \code{tdat} following the method specified by \code{bwtype}. The bandwidths
  \code{h} can be calculated from other functions in this package.
}
\description{
This is the description of the function \code{kde_h}.
}
\details{
These are the details of the function.
}
\section{Warning These can be some warnings.}{

}
\examples{

}

