% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdke_utils.R
\name{ruleOfThumb}
\alias{ruleOfThumb}
\title{Rule-of-thumb bandwidth}
\usage{
ruleOfThumb(tdat)
}
\arguments{
\item{tdat}{array containing the training dataset.}
}
\value{
a number corresponding to the standard bandwidth
}
\description{
\code{ruleOfThumb} calculates the rule-of-thumb bandwidth according to the
standard formula (eg. p23 of np package documentation).
}
\details{
These are the details of the function.
}
\examples{
ruleOfThumb(runif(100,0,1))
}

