% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdke_utils.R
\name{sample_dist_to_knn}
\alias{sample_dist_to_knn}
\title{calculates vector of nearest of nearest neighbors}
\usage{
sample_dist_to_knn(sampleVect, k)
}
\arguments{
\item{sampleVect}{vector containing sample datapoints.}

\item{k}{integer specifying the order of the neighbor.}
}
\value{
a vector of distances, where the \code{i}th element contains the
  distance between the \code{i}th element of \code{sampleVect} and its
  \code{k}th neighbor.
}
\description{
\code{sample_dist_to_knn} calculates kth-nearest-neighbor distances from a
sample.
}
\details{
These are the details of the function.
}
\examples{
sample_dist_to_knn(1:4,3)
}

