% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdke_utils.R
\name{vectToArray}
\alias{vectToArray}
\title{change vector to array}
\usage{
vectToArray(vect)
}
\arguments{
\item{vect}{vector}
}
\value{
an array with dimensions (1,length(vect))
}
\description{
change vector to array
}
\details{
Transforms a dataset in the form of a vector in a dataset in the form of an
array
}
\examples{
vectToArray(1:10)
}

