% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balloon_h.R
\name{bw_balloon_LQ}
\alias{bw_balloon_LQ}
\title{Loftsgaarden-Quesenberry kth nearest neighbor bandwidths}
\usage{
bw_balloon_LQ(tdat, edat, k)
}
\arguments{
\item{tdat}{a vector containing the training dataset.}

\item{edat}{a vector containing the evaluation data points.}

\item{k}{specifying the order of the neighbor.}
}
\value{
a vector of bandwidths  (one bandwidth per evaluation data point)
}
\description{
code{bw_ballon_LQ} looks for the bandwidth following the
Loftsgaarden-Quesenberry algorithm The bandwidth depends on the point of
estimation.
}
\details{
These are the details of the function.
}
\section{Warning These can be some warnings.}{

}
\examples{
bw_balloon_LQ(tdat=runif(n=10,min=0,max=2),edat=c(0.1,0.5,1),k=3)
}

