% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdke_utils.R
\name{dist_to_knn}
\alias{dist_to_knn}
\title{Distance to kth nearest neighbor}
\usage{
dist_to_knn(x, nghDat, k)
}
\arguments{
\item{x}{number (univariate) or vector (multivariate) corresponding to the
reference point.}

\item{nghDat}{vector (univariate) or array (multivariate) of numbers
corresponding to the neighbors.}

\item{k}{integer specifying the order of the neighbor.}
}
\value{
The distance between \code{x} and its \code{k}th nearest neighbor in
  \code{nghDat}.
}
\description{
\code{dist_to_knn} calculates the distance between one point and its kth
nearest neighbor in a vector
}
\details{
These are the details of the function.
}
\examples{
dist_to_knn(x = 1.4, nghDat = 0:10, k = 4)
}

