% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdke_utils.R
\name{sample_dist_to_knn}
\alias{sample_dist_to_knn}
\title{calculates vector of nearest of nearest neighbors}
\usage{
sample_dist_to_knn(sampleDat, k)
}
\arguments{
\item{sampleDat}{vector containing sample datapoints.}

\item{k}{integer specifying the order of the neighbor.}
}
\value{
a vector of distances, where the \code{i}th element contains the
  distance between the \code{i}th element of \code{sampleDat} and its
  \code{k}th neighbor.
}
\description{
\code{sample_dist_to_knn} calculates kth-nearest-neighbor distances from a
sample.
}
\details{
These are the details of the function.
}
\examples{
sample_dist_to_knn(sampleDat = 1:4,k = 3)
sample_dist_to_knn(sampleDat = array(1:12,dim = c(3,4)),k = 1)
}

