% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vkde.R
\name{sgk}
\alias{sgk}
\title{Single Gaussian Kernel}
\usage{
sgk(t, e, h)
}
\arguments{
\item{t}{a number corresponding to the training data point.}

\item{e}{a number or numeric vector corresponding to the evaluation data
point.}

\item{h}{the width of the kernel.}
}
\value{
The evaluation at location \code{e} of the Gaussian kernel centered
  at \code{t} and with width \code{h}.
}
\description{
\code{sgk} evaluates a kernel with a given center and bandwidth.
}
\details{
These are the details of the function.
}
\examples{
singleKde <- sgk(0,seq(from=-5,to=5,by=0.1),1)
plot(singleKde$kpdf$x,singleKde$kpdf$y,
   type='l',
   xlab = 'x',ylab='kernel density estimate')
   axis(1,singleKde$t,labels = NA,lwd.ticks = 2,col.ticks = 'red',tck=0.1)
}

