% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdke_utils.R
\name{vectToArray}
\alias{vectToArray}
\title{change vector to array}
\usage{
vectToArray(vect)
}
\arguments{
\item{vect}{vector}
}
\value{
an array with dimensions (1,length(vect)
}
\description{
change vector to array
}
\details{
change vector to array
}
\examples{
vectToArray(1:10)
}

